refactor: TeamInfo의 메인페이지 변경
